+++ src/main/java/ru/yandex/practicum/filmorate/dao/film/FilmStorage.java
@@ -1,15 +1,10 @@
 package ru.yandex.practicum.filmorate.dao.film;
 
-import ru.yandex.practicum.filmorate.model.Director;
-import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.model.Genre;
-import ru.yandex.practicum.filmorate.model.RatingMpa;
-
-import java.util.Collection;
-import java.util.List;
+import ru.yandex.practicum.filmorate.model.*;
+import java.util.*;
 
 public interface FilmStorage {
-    Film getFilmsById(Integer id);
+    Film getFilmsById(Long id);
 
     Collection<Film> getAllFilms();
 
@@ -27,13 +22,13 @@ public interface FilmStorage {
 
     Film updateFilm(Film film);
 
-    void deleteFilm(Integer id);
+    void deleteFilm(Long id);
 
-    Film putLike(Integer id, Long userId);
+    Film putLike(Long id, Long userId);
 
-    Film deleteLike(Integer id, Long userId);
+    Film deleteLike(Long id, Long userId);
 
-    boolean isExistsIdFilm(Integer filmId);
+    boolean isExistsIdFilm(Long filmId);
 
     Collection<Film> getCommonFilms(Long userId, Long friendId);
