+++ src/main/java/ru/yandex/practicum/filmorate/mapper/GenreFromFilmMapper.java
@@ -3,19 +3,14 @@ package ru.yandex.practicum.filmorate.mapper;
 import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.ResultSetExtractor;
 import ru.yandex.practicum.filmorate.model.Genre;
+import java.sql.*;
+import java.util.*;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-
-public class GenreFromFilmMapper implements ResultSetExtractor<Map<Integer, List<Genre>>> {
-    private final Map<Integer, List<Genre>> genresOfFilms = new HashMap<>();
+public class GenreFromFilmMapper implements ResultSetExtractor<Map<Long, List<Genre>>> {
+    private final Map<Long, List<Genre>> genresOfFilms = new HashMap<>();
 
     @Override
-    public Map<Integer, List<Genre>> extractData(ResultSet rs) throws DataAccessException, SQLException {
+    public Map<Long, List<Genre>> extractData(ResultSet rs) throws DataAccessException, SQLException {
         while (rs.next()) {
             if (isFilmGenresInMap(rs)) {
                 addToExistingList(rs);
@@ -27,18 +22,18 @@ public class GenreFromFilmMapper implements ResultSetExtractor<Map<Integer, List
     }
 
     private boolean isFilmGenresInMap(ResultSet rs) throws SQLException {
-        return genresOfFilms.containsKey(rs.getInt("film_id"));
+        return genresOfFilms.containsKey(rs.getLong("film_id"));
     }
 
     private void addToExistingList(ResultSet rs) throws SQLException {
-        List<Genre> genres = genresOfFilms.get(rs.getInt("film_id"));
+        List<Genre> genres = genresOfFilms.get(rs.getLong("film_id"));
         genres.add(new Genre(rs.getInt("genre_id"), rs.getString("name")));
-        genresOfFilms.put(rs.getInt("film_id"), genres);
+        genresOfFilms.put(rs.getLong("film_id"), genres);
     }
 
     private void addNewList(ResultSet rs) throws SQLException {
         List<Genre> genresOfOneFilm = new LinkedList<>();
         genresOfOneFilm.add(new Genre(rs.getInt("genre_id"), rs.getString("name")));
-        genresOfFilms.put(rs.getInt("film_id"), genresOfOneFilm);
+        genresOfFilms.put(rs.getLong("film_id"), genresOfOneFilm);
     }
 }
